refactor(render): consolidate doctor/health-check with statusPanel (#1404)
Extract healthCheckRows() and healthCheckSummary() into status-panel.ts
so both doctor surfaces (CLI command and /doctor slash command) share
the same rendering logic instead of duplicating it.
