+++ src/cli/commands/doctor.ts
@@ -1,5 +1,5 @@
 import { Command } from 'commander';
-import { palette } from '../palette.js';
+import { healthCheckRows, healthCheckSummary } from '../render/status-panel.js';
 import { runDoctorChecks } from './doctor-checks.js';
 
 export function registerDoctorCommand(program: Command): void {
@@ -19,30 +19,12 @@ export function registerDoctorCommand(program: Command): void {
       if (options.format === 'json') {
         console.log(JSON.stringify({ checks, summary }, null, 2));
       } else {
-        checks.forEach((check) => {
-          let icon: string;
-          if (check.state === 'pass') {
-            icon = palette.success('✓');
-          } else if (check.state === 'warn') {
-            icon = palette.warning('⚠');
-          } else {
-            icon = palette.error('✗');
+        for (const check of checks) {
+          for (const line of healthCheckRows(check)) {
+            console.log(line);
           }
-
-          let line = `${icon} ${check.name}`;
-          if (check.detail) {
-            line += ` — ${check.detail}`;
-          }
-          console.log(line);
-
-          if (check.state === 'fail' && check.fix) {
-            console.log(`  Fix: ${check.fix}`);
-          }
-        });
-
-        console.log(
-          `\nSummary: ${summary.passed} passed, ${summary.warned} warned, ${summary.failed} failed`,
-        );
+        }
+        console.log('\n' + healthCheckSummary(checks));
       }
 
       process.exit(summary.failed > 0 ? 1 : 0);
